+++ airbyte-server/src/main/java/io/airbyte/server/errors/IdNotFoundExceptionMapper.java
@@ -2,8 +2,10 @@
  * Copyright (c) 2022 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.commons.server.errors;
+package io.airbyte.server.errors;
 
+import io.airbyte.commons.server.errors.IdNotFoundKnownException;
+import io.airbyte.commons.server.errors.KnownException;
 import io.micronaut.context.annotation.Requires;
 import io.micronaut.http.HttpRequest;
 import io.micronaut.http.HttpResponse;
@@ -12,20 +14,19 @@ import io.micronaut.http.MediaType;
 import io.micronaut.http.annotation.Produces;
 import io.micronaut.http.server.exceptions.ExceptionHandler;
 import jakarta.inject.Singleton;
-import javax.ws.rs.NotFoundException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 @Produces
 @Singleton
-@Requires(classes = NotFoundException.class)
-public class NotFoundExceptionHandler implements ExceptionHandler<NotFoundException, HttpResponse> {
+@Requires(classes = IdNotFoundKnownException.class)
+public class IdNotFoundExceptionMapper implements ExceptionHandler<IdNotFoundKnownException, HttpResponse> {
 
-  private static final Logger LOGGER = LoggerFactory.getLogger(NotFoundExceptionHandler.class);
+  private static final Logger LOGGER = LoggerFactory.getLogger(IdNotFoundExceptionMapper.class);
 
   @Override
-  public HttpResponse handle(final HttpRequest request, final NotFoundException exception) {
-    final IdNotFoundKnownException idnf = new IdNotFoundKnownException("Object not found. " + exception.getMessage(), exception);
+  public HttpResponse handle(final HttpRequest request, final IdNotFoundKnownException exception) {
+    final IdNotFoundKnownException idnf = new IdNotFoundKnownException("Id not found: " + exception.getMessage(), exception);
     LOGGER.error("Not found exception", idnf.getNotFoundKnownExceptionInfo());
 
     return HttpResponse.status(HttpStatus.NOT_FOUND)
