+++ airbyte-server/src/main/java/io/airbyte/server/errors/InvalidJsonExceptionHandler.java
@@ -2,9 +2,10 @@
  * Copyright (c) 2022 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.commons.server.errors;
+package io.airbyte.server.errors;
 
-import com.fasterxml.jackson.databind.JsonMappingException;
+import com.fasterxml.jackson.core.JsonParseException;
+import io.airbyte.commons.server.errors.KnownException;
 import io.micronaut.context.annotation.Requires;
 import io.micronaut.http.HttpRequest;
 import io.micronaut.http.HttpResponse;
@@ -16,14 +17,14 @@ import jakarta.inject.Singleton;
 
 @Produces
 @Singleton
-@Requires(classes = JsonMappingException.class)
-public class InvalidJsonInputExceptionHandler implements ExceptionHandler<JsonMappingException, HttpResponse> {
+@Requires(classes = JsonParseException.class)
+public class InvalidJsonExceptionHandler implements ExceptionHandler<JsonParseException, HttpResponse> {
 
   @Override
-  public HttpResponse handle(final HttpRequest request, final JsonMappingException exception) {
+  public HttpResponse handle(final HttpRequest request, final JsonParseException exception) {
     return HttpResponse.status(HttpStatus.UNPROCESSABLE_ENTITY)
-        .body(KnownException.infoFromThrowableWithMessage(exception,
-            "Invalid json input. " + exception.getMessage() + " " + exception.getOriginalMessage()))
+        .body(
+            KnownException.infoFromThrowableWithMessage(exception, "Invalid json. " + exception.getMessage() + " " + exception.getOriginalMessage()))
         .contentType(MediaType.APPLICATION_JSON_TYPE);
   }
