+++ airbyte-server/src/main/java/io/airbyte/server/errors/InvalidJsonExceptionMapper.java
@@ -2,9 +2,10 @@
  * Copyright (c) 2022 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.commons.server.errors;
+package io.airbyte.server.errors;
 
 import com.fasterxml.jackson.core.JsonParseException;
+import io.airbyte.commons.server.errors.KnownException;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.ext.ExceptionMapper;
 import javax.ws.rs.ext.Provider;
