+++ airbyte-server/src/main/java/io/airbyte/server/errors/UncaughtExceptionMapper.java
@@ -2,10 +2,11 @@
  * Copyright (c) 2022 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.commons.server.errors;
+package io.airbyte.server.errors;
 
 import io.airbyte.api.model.generated.KnownExceptionInfo;
 import io.airbyte.commons.json.Jsons;
+import io.airbyte.commons.server.errors.KnownException;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.ext.ExceptionMapper;
 import javax.ws.rs.ext.Provider;
